サービス名変更 (#57)
Pull request overview
prsk-pr-labeling を廃止し、新しい Composite Action prsk-cheering（ラベル付与＋応援コメント投稿）へサービス名/構成を移行するPRです。README/ガイド類とDependabot設定も新ディレクトリに追従させています。

Changes:

prsk-pr-labeling の Action/ロジックを削除し、prsk-cheering に再配置（ワークフロー本体・コメント/ラベルロジック分割・定数JSON追加）
.github/workflows/prsk-cheering.yml と .github/dependabot.yml を新Actionに対応
README/AGENTSの記載更新、およびPR/Issueテンプレートの削除
