+++ AGENTS.md
@@ -13,12 +13,19 @@
 ├── AGENTS.md # エージェント向けの補助ガイド
 ├── CLAUDE.md # プロジェクトの基本方針
 ├── README.md # リポジトリの概要と使い方
+├── LICENSE
+├── VERSION
 ├── prsk-dependabot/ # Dependabot関連のComposite Action
 │   ├── action.yml # Action定義
 │   ├── prsk-dependabot.constants.json # 依存関係更新の設定値
 │   └── prsk-pr-rename.js # PRタイトル調整ロジック
-└── prsk-pr-labeling/ # PRラベル付けのComposite Action
-    ├── action.yml # Action定義
-    ├── prsk-labeling-logic.js # ラベル判定ロジック
-    └── prsk-yell-label.constants.json # ラベル関連の設定値
+├── prsk-cheering/ # チアリング関連のComposite Action
+│   ├── action.yml # Action定義
+│   ├── workflow.js # メインワークフロー（オーケストレーション）
+│   ├── prsk-labeling-logic.js # PRラベル付与ロジック
+│   ├── prsk-comments-logic.js # PR応援コメント投稿ロジック
+│   ├── utils.js # 共通ユーティリティ
+│   └── prsk-yell-label.constants.json # ラベルや応援コメントの定義
+└── .github/ # GitHub ワークフロー
+    └── workflows/
 ```
