--- prsk-pr-labeling/action.yml
@@ -1,21 +1,0 @@
-name: 'Project SEKAI PR Labeling'
-description: 'Set random Project SEKAI character labels to pull requests'
-author: 'Narumikr'
-
-inputs:
-  github-token:
-    description: 'GitHub token for API access'
-    required: true
-    default: ${{ github.token }}
-
-runs:
-  using: 'composite'
-  steps:
-    - name: Create and add character label
-      uses: actions/github-script@v8
-      with:
-        github-token: ${{ inputs.github-token }}
-        script: |
-          const { handlePrLabeling } = require('${{ github.action_path }}/prsk-labeling-logic.js');
-
-          await handlePrLabeling(github, context, '${{ github.action_path }}');
